Refactoring  - Refactored the MNIST quantized CNN to simplify the implementation    of other datasets  - Quantization functions are now in common src directory  - Also moved other common functions which can be reused  - Tidy up files and dirs
