+++ src/variable_initialization.py
@@ -1,0 +1,11 @@
+import tensorflow as tf
+
+# Initialize weight variable as random tensor with stddev 0.1
+def weight_variable(shape):
+  initial = tf.truncated_normal(shape, stddev=0.1)
+  return tf.Variable(initial)
+
+# Initialize bias as constant 0.1
+def bias_variable(shape):
+  initial = tf.constant(0.1, shape=shape)
+  return tf.Variable(initial)
